moved Project, Subject, Visit and Session into separate module
