+++ arcana/archive/__init__.py
@@ -1,2 +1,3 @@
 from .xnat import XnatArchive
 from .local import LocalArchive
+from .tree import Project, Subject, Visit, Session
